requirements.txt
# Solana & Web3
solana>=0.30.2
solders>=0.18.0
anchorpy>=0.19.0
base58>=2.1.1

# HTTP & API
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# CLI & UI
click>=8.1.7
rich>=13.7.0
pyfiglet>=1.0.2

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Notifications
plyer>=2.1.0
pygame>=2.5.2

# Utilities
colorama>=0.4.6
python-dateutil>=2.8.2
pytz>=2023.3

# Crypto
cryptography>=41.0.7
#nacl>=1.5.0
nacl
